Refactor InventoryRepositoriesCommand into partial classes
